requirements.txt
python-dotenv
openai
fastapi
uvicorn
PyJWT
requests
cryptography
jinja2